requirements.txt
Django==1.8.6
djangorestframework==3.4.6
djangorestframework-jwt==1.8.0
drfdocs==0.0.11
PyJWT==1.4.2
wheel==0.24.0